Move 'matchOwnerReferenceWithResources' unit tests from 'pkg/logs' to 'pkg/kclient'
Those tests are specific to the Kubernetes implementation.
